+++ src/unix/misc_unix.c
@@ -59,3 +59,8 @@ int ev__translate_sys_error(int syserr)
 
     return syserr;
 }
+
+void ev_library_shutdown(void)
+{
+    // Do nothing
+}